+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Damage/Tabs.lua
@@ -22,6 +22,7 @@ function MyDungeonsBook:DamageFrame_CreateTabButtonsFrame(parentFrame)
 		{value = "damageDoneByPartyMembers", text = L["All damage done"]},
 		{value = "damageDoneToUnits", text = L["Damage Done To Units"]},
 		{value = "friendlyFire", text = L["Friendly Fire"]},
+		{value = "enemiesFriendlyFire", text = L["Enemies Friendly Fire"]},
 	});
 	tabs:SetCallback("OnGroupSelected", function (container, _, tabId)
 		container:ReleaseChildren();
@@ -36,6 +37,9 @@ function MyDungeonsBook:DamageFrame_CreateTabButtonsFrame(parentFrame)
 		end
 		if (tabId == "friendlyFire") then
 			self.friendlyFireFrame = self:FriendlyFireFrame_Create(container, self.activeChallengeId);
+		end
+		if (tabId == "enemiesFriendlyFire") then
+			self.enemiesFriendlyFireFrame = self:EnemiesFriendlyFireFrame_Create(container, self.activeChallengeId);
 		end
 	end);
 	tabs:SetHeight(586);